+++ eminus/__init__.py
@@ -9,6 +9,7 @@ Minimal usage example to do a DFT calculation for helium::
 """
 from . import config
 from .atoms import Atoms
+from .cell import Cell
 from .dft import get_epsilon, get_psi
 from .io import (
     read,
@@ -27,7 +28,7 @@ from .logger import log
 from .scf import RSCF, SCF, USCF
 from .version import __version__, info
 
-__all__ = ['config', 'Atoms', 'get_epsilon', 'get_psi', 'info', 'log', 'read', 'read_cube',
+__all__ = ['config', 'Atoms', 'Cell', 'get_epsilon', 'get_psi', 'info', 'log', 'read', 'read_cube',
            'read_json', 'read_traj', 'read_xyz', 'RSCF', 'SCF', 'USCF', 'write', 'write_cube',
            'write_json', 'write_pdb', 'write_traj', 'write_xyz', '__version__']